feat: swiftui example draft - need minor design changes
