--- examples/swiftui/SwiftUI Example/Constants.swift
@@ -1,18 +1,0 @@
-//
-//  Constants.swift
-//  SwiftUI Example
-//
-//  Created by 최성혁 on 2023/07/24.
-//
-
-import Foundation
-
-struct Constants {
-    struct Layout {
-
-    }
-
-    struct Color {
-
-    }
-}
