+++ examples/swiftui/SwiftUI Example/Helpers/ReplayButton.swift
@@ -8,16 +8,15 @@
 import SwiftUI
 
 struct ReplayButton: View {
-    @Binding var roomId: String
-    @Binding var accessToken: String
-    @Binding var query: String
-    @Binding var isAlertOn: Bool
+    private var onTap: () -> (Void)
+    
+    init(onTap: @escaping () -> (Void)) {
+        self.onTap = onTap
+    }
 
     var body: some View {
         Button(action: {
-            if roomId == "" || accessToken == "" {
-                isAlertOn = true
-            }
+            onTap()
         }) {
             Text("Replay")
                 .font(Font.custom("Pretendard", size: 14).weight(.medium))
@@ -38,10 +37,7 @@ struct ReplayButton: View {
 }
 
 struct ReplayButton_Previews: PreviewProvider {
-    @State static private var text = ""
-    @State static private var isAlertOn = false
-
     static var previews: some View {
-        ReplayButton(roomId: $text, accessToken: $text, query: $text, isAlertOn: $isAlertOn)
+        ReplayButton(onTap: {})
     }
 }
