+++ examples/swiftui/SwiftUI Example/SwiftUI_ExampleApp.swift
@@ -4,7 +4,7 @@ import SwiftUI
 struct SwiftUI_ExampleApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            HomeView()
         }
     }
 }
